fix(name): ignore pyutilib name overwrite
Upon creation of a plugin instance, `pyutilib` attempts to assign a
unique name to the instance. However, in practice, all MicroDrop plugins
override the name assigned by `pyutilib` and replace it with the value
of the `plugin_name` attribute of the respective plugin.

In this commit, ignore attempts by `pyutilib` to overwrite the plugin
`name` attribute.